third_party/blink/web_tests/plugins/plugin-reload-data.html

<!DOCTYPE html>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();

        onload = () => {
            var object = document.querySelector("object");
            object.setAttribute("data", "data:text/plain,second");
            object.addEventListener("message", () => { testRunner.notifyDone(); });
        };
    } else {
        document.write("Needs to be run as a layout test.");
    }
</script>
<object data="data:text/plain,first" type="application/x-blink-test-plugin" logargs="true"></object>
