third_party/blink/web_tests/plugins/plugin-reload-data-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: plugin args:
CONSOLE MESSAGE: Blink Test Plugin:   name = data, value = data:text/plain,first
CONSOLE MESSAGE: Blink Test Plugin:   name = type, value = application/x-blink-test-plugin
CONSOLE MESSAGE: Blink Test Plugin:   name = logargs, value = true
CONSOLE MESSAGE: Blink Test Plugin:   name = src, value = data:text/plain,first
CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: plugin args:
CONSOLE MESSAGE: Blink Test Plugin:   name = data, value = data:text/plain,second
CONSOLE MESSAGE: Blink Test Plugin:   name = type, value = application/x-blink-test-plugin
CONSOLE MESSAGE: Blink Test Plugin:   name = logargs, value = true
CONSOLE MESSAGE: Blink Test Plugin:   name = src, value = data:text/plain,second
